remove yarnconfig. Move backend configs to backend.rs files
