+++ src/prelude.rs
@@ -2,19 +2,19 @@ pub use crate::backends::Backend;
 pub use crate::backends::all::{AnyBackend, PackageIds, Packages, RawPackageIds, RawPackages};
 pub(crate) use crate::backends::apply_backends;
 pub use crate::backends::apt::{Apt, AptOptions};
-pub use crate::backends::arch::{Arch, ArchOptions};
+pub use crate::backends::arch::{Arch, ArchConfig, ArchOptions};
 pub use crate::backends::brew::{Brew, BrewOptions};
 pub use crate::backends::bun::{Bun, BunOptions};
-pub use crate::backends::cargo::{Cargo, CargoOptions};
+pub use crate::backends::cargo::{Cargo, CargoConfig, CargoOptions};
 pub use crate::backends::dnf::{Dnf, DnfOptions};
-pub use crate::backends::flatpak::{Flatpak, FlatpakOptions};
+pub use crate::backends::flatpak::{Flatpak, FlatpakConfig, FlatpakOptions};
 pub use crate::backends::npm::{Npm, NpmOptions};
 pub use crate::backends::pipx::{Pipx, PipxOptions};
 pub use crate::backends::pnpm::{Pnpm, PnpmOptions};
 pub use crate::backends::scoop::{Scoop, ScoopGetOptions};
 pub use crate::backends::snap::{Snap, SnapOptions};
 pub use crate::backends::uv::{Uv, UvOptions};
-pub use crate::backends::vscode::{VsCode, VsCodeOptions};
+pub use crate::backends::vscode::{VsCode, VsCodeConfig, VsCodeOptions};
 pub use crate::backends::winget::{WinGet, WinGetOptions};
 pub use crate::backends::xbps::{Xbps, XbpsOptions};
 pub use crate::backends::yarn::{Yarn, YarnOptions};
@@ -23,8 +23,6 @@ pub use crate::cli::{
     SyncCommand, UninstallCommand, UnmanagedCommand, UpdateAllCommand, UpdateCommand,
 };
 pub use crate::cmd::Perms;
-pub use crate::config::{
-    ArchConfig, ArchPackageManager, CargoConfig, Config, FlatpakConfig, VsCodeConfig, VsCodeVariant,
-};
+pub use crate::config::{ArchPackageManager, Config, VsCodeVariant};
 pub use crate::groups::Groups;
 pub use crate::package::{Hooks, Package};
